Feat: OrderController, Customer free to access and create order, getAllOrders, Only staff can change the status of the order

+++ src/main/java/com/example/SelfOrderingRestaurant/Dto/Request/DishRequestDTO/DishRequestDTO.java
@@ -13,7 +13,7 @@ import java.math.BigDecimal;
 @NoArgsConstructor
 public class DishRequestDTO {
     private String name;
-    private Long price;
+    private BigDecimal price;
     private Integer categoryId;
     private String description;
     private String image;
